+++ src/test/java/com/tdunning/math/stats/GroupTreeTest.java
@@ -17,19 +17,17 @@
 
 package com.tdunning.math.stats;
 
-import com.google.common.collect.Lists;
+import java.util.Collections;
+import java.util.Iterator;
+import java.util.List;
+
 import org.apache.mahout.common.RandomUtils;
 import org.junit.Before;
 import org.junit.Test;
 
-import java.util.Collections;
-import java.util.Iterator;
-import java.util.List;
-import java.util.Random;
+import com.google.common.collect.Lists;
 
-import static org.junit.Assert.*;
-
-public class GroupTreeTest {
+public class GroupTreeTest extends AbstractTest {
     @Test
     public void testSimpleAdds() {
         GroupTree x = new GroupTree();
@@ -182,11 +180,10 @@ public class GroupTreeTest {
 
     @Test
     public void testRandomRebalance() {
-        Random gen = RandomUtils.getRandom();
         GroupTree x = new GroupTree();
         List<Double> y = Lists.newArrayList();
         for (int i = 0; i < 1000; i++) {
-            double v = gen.nextDouble();
+            double v = randomDouble();
             x.add(new Centroid(v));
             y.add(v);
             x.checkBalance();
@@ -200,7 +197,7 @@ public class GroupTreeTest {
         }
 
         for (int j = 0; j < 100; j++) {
-            double v = y.get(gen.nextInt(y.size()));
+            double v = y.get(randomInt(y.size() - 1));
             y.remove(v);
             x.remove(x.floor(new Centroid(v)));
         }
