+++ src/test/java/com/tdunning/math/stats/TDigestTest.java
@@ -17,41 +17,46 @@
 
 package com.tdunning.math.stats;
 
-import com.google.common.collect.Iterables;
-import com.google.common.collect.Lists;
-
-import org.apache.mahout.common.RandomUtils;
-import org.apache.mahout.math.jet.random.AbstractContinousDistribution;
-import org.junit.AfterClass;
-import org.junit.BeforeClass;
-import org.junit.Test;
-
 import java.io.File;
 import java.io.FileNotFoundException;
+import java.io.FileOutputStream;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.io.PrintWriter;
 import java.io.StringWriter;
+import java.nio.ByteBuffer;
 import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.Collection;
 import java.util.Collections;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Random;
 import java.util.concurrent.Callable;
 import java.util.concurrent.ExecutionException;
+import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.concurrent.Future;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertTrue;
-import static org.junit.Assert.fail;
+import org.apache.mahout.math.jet.random.AbstractContinousDistribution;
+import org.apache.mahout.math.jet.random.Gamma;
+import org.apache.mahout.math.jet.random.Normal;
+import org.apache.mahout.math.jet.random.Uniform;
+import org.junit.AfterClass;
+import org.junit.BeforeClass;
+import org.junit.Ignore;
+import org.junit.Test;
+
+import com.clearspring.analytics.stream.quantile.QDigest;
+import com.google.common.collect.Iterables;
+import com.google.common.collect.Lists;
 
 
 /**
- * Common test methods for TDigests
+ * Base test case for TDigests, just extend this class and implement the abstract methods.
  */
-public class TDigestTest {
+@Ignore
+public abstract class TDigestTest extends AbstractTest {
     protected static PrintWriter sizeDump;
     protected static PrintWriter errorDump;
     protected static PrintWriter deviationDump;
@@ -75,9 +80,43 @@ public class TDigestTest {
         deviationDump.close();
     }
 
+    public interface DigestFactory {
+        TDigest create();
+    }
 
-    protected void merge(final DigestFactory<? extends TDigest> factory) throws FileNotFoundException, InterruptedException, ExecutionException {
-        final Random gen0 = RandomUtils.getRandom();
+    protected abstract DigestFactory factory(double compression);
+
+    protected final DigestFactory factory() {
+        return factory(100);
+    }
+
+    protected abstract TDigest fromBytes(ByteBuffer bytes);
+
+    private static TDigest merge(Iterable<TDigest> subData, Random gen, TDigest r) {
+        List<Centroid> centroids = new ArrayList<Centroid>();
+        boolean recordAll = false;
+        for (TDigest digest : subData) {
+            for (Centroid centroid : digest.centroids()) {
+                centroids.add(centroid);
+            }
+            recordAll |= digest.isRecording();
+        }
+        Collections.shuffle(centroids, gen);
+        if (recordAll) {
+            r.recordAllData();
+        }
+
+        for (Centroid c : centroids) {
+            if (r.isRecording()) {
+                // TODO should do something better here.
+            }
+            ((AbstractTDigest) r).add(c.mean(), c.count(), c);
+        }
+        return r;
+    }
+
+    protected void merge(final DigestFactory factory) throws FileNotFoundException, InterruptedException, ExecutionException {
+        final Random gen0 = getRandom();
         PrintWriter out = new PrintWriter(new File("merge.tsv"));
         out.printf("type\tparts\tq\te0\te1\te2\te2.rel\n");
 
@@ -130,7 +169,7 @@ public class TDigestTest {
                         }
 
                         // now merge the sub-digests
-                        TDigest dist2 = AbstractTDigest.merge(subs, gen, factory.create());
+                        TDigest dist2 = merge(subs, gen, factory.create());
 
                         // verify the merged result has the right data
                         List<Double> data3 = Lists.newArrayList();
@@ -171,24 +210,30 @@ public class TDigestTest {
             });
         }
 
-        for (Future<String> result : Executors.newFixedThreadPool(20).invokeAll(tasks)) {
+        ExecutorService executor = Executors.newFixedThreadPool(20);
+        for (Future<String> result : executor.invokeAll(tasks)) {
             out.write(result.get());
         }
+        executor.shutdown();
         out.close();
     }
 
-    protected void singleValue(TDigest digest) {
-        final double value = RandomUtils.getRandom().nextDouble() * 1000;
+    @Test
+    public void testSingleValue() {
+        final TDigest digest = factory().create();
+        final double value = getRandom().nextDouble() * 1000;
         digest.add(value);
-        final double q = RandomUtils.getRandom().nextDouble();
+        final double q = getRandom().nextDouble();
         for (double qValue : new double[] {0, q, 1}) {
             assertEquals(value, digest.quantile(qValue), 0.001f);
         }
     }
 
-    protected void fewValues(TDigest digest) {
+    @Test
+    public void testFewValues() {
         // When there are few values in the tree, quantiles should be exact
-        final Random r = RandomUtils.getRandom();
+        final TDigest digest = factory().create();
+        final Random r = getRandom();
         final int length = r.nextInt(10);
         final List<Double> values = new ArrayList<Double>();
         for (int i = 0; i < length; ++i) {
@@ -247,7 +292,7 @@ public class TDigestTest {
      * @param tag           Label for the output lines
      * @param recordAllData True if the internal histogrammer should be set up to record all data it sees for
      */
-    protected void runTest(DigestFactory<? extends TDigest> factory, AbstractContinousDistribution gen, double sizeGuide, double[] qValues, String tag, boolean recordAllData) {
+    protected void runTest(DigestFactory factory, AbstractContinousDistribution gen, double sizeGuide, double[] qValues, String tag, boolean recordAllData) {
         TDigest dist = factory.create();
         if (recordAllData) {
             dist.recordAllData();
@@ -328,13 +373,17 @@ public class TDigestTest {
         }
     }
 
-    protected void empty(TDigest digest) {
-        final double q = RandomUtils.getRandom().nextDouble();
+    @Test
+    public void testEmpty() {
+        final TDigest digest = factory().create();
+        final double q = getRandom().nextDouble();
         assertTrue(Double.isNaN(digest.quantile(q)));
     }
 
-    protected void moreThan2BValues(TDigest digest) {
-        Random gen = RandomUtils.getRandom();
+    @Test
+    public void testMoreThan2BValues() {
+        final TDigest digest = factory().create();
+        Random gen = getRandom();
         for (int i = 0; i < 1000; ++i) {
             final double next = gen.nextDouble();
             digest.add(next);
@@ -357,21 +406,9 @@ public class TDigestTest {
     }
 
     @Test
-    public void testMergeEmpty() {
-        final Random gen0 = RandomUtils.getRandom();
-        List<TDigest> subData = new ArrayList();
-        subData.add(new TreeDigest(10));
-        TreeDigest foo = new TreeDigest(10);
-        AbstractTDigest.merge(subData, gen0, foo);
-        empty(foo);
-    }
-
-    public interface DigestFactory<T extends TDigest> {
-        T create();
-    }
-
-    protected void sorted(TDigest digest) {
-        Random gen = RandomUtils.getRandom();
+    public void testSorted() {
+        final TDigest digest = factory().create();
+        Random gen = getRandom();
         for (int i = 0; i < 10000; ++i) {
             digest.add(gen.nextDouble(), 1 + gen.nextInt(10));
         }
@@ -384,8 +421,10 @@ public class TDigestTest {
         }
     }
 
-    protected void nan(TDigest digest) {
-        Random gen = RandomUtils.getRandom();
+    @Test
+    public void testNaN() {
+        final TDigest digest = factory().create();
+        Random gen = getRandom();
         final int iters = gen.nextInt(100);
         for (int i = 0; i < iters; ++i) {
             digest.add(gen.nextDouble(), 1 + gen.nextInt(10));
@@ -400,6 +439,415 @@ public class TDigestTest {
             fail("NaN should be an illegal argument");
         } catch (IllegalArgumentException e) {
             // expected
+        }
+    }
+
+    @Test
+    public void testUniform() {
+        Random gen = getRandom();
+        for (int i = 0; i < repeats(); i++) {
+            runTest(factory(), new Uniform(0, 1, gen), 100,
+                    new double[]{0.001, 0.01, 0.1, 0.5, 0.9, 0.99, 0.999},
+                    "uniform", true);
+        }
+    }
+
+    @Test
+    public void testGamma() {
+        // this Gamma distribution is very heavily skewed.  The 0.1%-ile is 6.07e-30 while
+        // the median is 0.006 and the 99.9th %-ile is 33.6 while the mean is 1.
+        // this severe skew means that we have to have positional accuracy that
+        // varies by over 11 orders of magnitude.
+        Random gen = getRandom();
+        for (int i = 0; i < repeats(); i++) {
+            runTest(factory(), new Gamma(0.1, 0.1, gen), 100,
+//                    new double[]{6.0730483624079e-30, 6.0730483624079e-20, 6.0730483627432e-10, 5.9339110446023e-03,
+//                            2.6615455373884e+00, 1.5884778179295e+01, 3.3636770117188e+01},
+                    new double[]{0.001, 0.01, 0.1, 0.5, 0.9, 0.99, 0.999},
+                    "gamma", true);
+        }
+    }
+
+    @Test
+    public void testNarrowNormal() {
+        // this mixture of a uniform and normal distribution has a very narrow peak which is centered
+        // near the median.  Our system should be scale invariant and work well regardless.
+        final Random gen = getRandom();
+        AbstractContinousDistribution mix = new AbstractContinousDistribution() {
+            AbstractContinousDistribution normal = new Normal(0, 1e-5, gen);
+            AbstractContinousDistribution uniform = new Uniform(-1, 1, gen);
+
+            @Override
+            public double nextDouble() {
+                double x;
+                if (gen.nextDouble() < 0.5) {
+                    x = uniform.nextDouble();
+                } else {
+                    x = normal.nextDouble();
+                }
+                return x;
+            }
+        };
+
+        for (int i = 0; i < repeats(); i++) {
+            runTest(factory(), mix, 100, new double[]{0.001, 0.01, 0.1, 0.3, 0.5, 0.7, 0.9, 0.99, 0.999}, "mixture", false);
+        }
+    }
+
+    @Test
+    public void testRepeatedValues() {
+        final Random gen = getRandom();
+
+        // 5% of samples will be 0 or 1.0.  10% for each of the values 0.1 through 0.9
+        AbstractContinousDistribution mix = new AbstractContinousDistribution() {
+            @Override
+            public double nextDouble() {
+                return Math.rint(gen.nextDouble() * 10) / 10.0;
+            }
+        };
+
+        TDigest dist = factory(1000).create();
+        List<Double> data = Lists.newArrayList();
+        for (int i1 = 0; i1 < 100000; i1++) {
+            double x = mix.nextDouble();
+            data.add(x);
+        }
+
+        long t0 = System.nanoTime();
+        for (double x : data) {
+            dist.add(x);
+        }
+
+        System.out.printf("# %fus per point\n", (System.nanoTime() - t0) * 1e-3 / 100000);
+        System.out.printf("# %d centroids\n", dist.centroids().size());
+
+        // I would be happier with 5x compression, but repeated values make things kind of weird
+        assertTrue("Summary is too large: " + dist.centroids().size(), dist.centroids().size() < 10 * (double) 1000);
+
+        // all quantiles should round to nearest actual value
+        for (int i = 0; i < 10; i++) {
+            double z = i / 10.0;
+            // we skip over troublesome points that are nearly halfway between
+            for (double delta : new double[]{0.01, 0.02, 0.03, 0.07, 0.08, 0.09}) {
+                double q = z + delta;
+                double cdf = dist.cdf(q);
+                // we also relax the tolerances for repeated values
+                assertEquals(String.format("z=%.1f, q = %.3f, cdf = %.3f", z, q, cdf), z + 0.05, cdf, 0.01);
+
+                double estimate = dist.quantile(q);
+                assertEquals(String.format("z=%.1f, q = %.3f, cdf = %.3f, estimate = %.3f", z, q, cdf, estimate), Math.rint(q * 10) / 10.0, estimate, 0.001);
+            }
+        }
+    }
+
+    @Test
+    public void testSequentialPoints() {
+        for (int i = 0; i < repeats(); i++) {
+            runTest(factory(), new AbstractContinousDistribution() {
+                double base = 0;
+
+                @Override
+                public double nextDouble() {
+                    base += Math.PI * 1e-5;
+                    return base;
+                }
+            }, 100, new double[]{0.001, 0.01, 0.1, 0.5, 0.9, 0.99, 0.999},
+                    "sequential", true);
+        }
+    }
+
+    @Test
+    public void testSerialization() {
+        Random gen = getRandom();
+        final double compression = randomDouble() * 100;
+        TDigest dist = factory(compression).create();
+        for (int i = 0; i < 100000; i++) {
+            double x = gen.nextDouble();
+            dist.add(x);
+        }
+        dist.compress();
+
+        ByteBuffer buf = ByteBuffer.allocate(20000);
+        dist.asBytes(buf);
+        assertTrue(buf.position() < 11000);
+        assertEquals(buf.position(), dist.byteSize());
+        buf.clear();
+
+        dist.asSmallBytes(buf);
+        assertTrue(buf.position() < 6000);
+        assertEquals(buf.position(), dist.smallByteSize());
+
+        System.out.printf("# big %d bytes\n", buf.position());
+
+        buf.flip();
+        TDigest dist2 = fromBytes(buf);
+        assertEquals(dist.centroids().size(), dist2.centroids().size());
+        assertEquals(dist.compression(), dist2.compression(), 0);
+        assertEquals(dist.size(), dist2.size());
+
+        for (double q = 0; q < 1; q += 0.01) {
+            assertEquals(dist.quantile(q), dist2.quantile(q), 1e-8);
+        }
+
+        Iterator<? extends Centroid> ix = dist2.centroids().iterator();
+        for (Centroid centroid : dist.centroids()) {
+            assertTrue(ix.hasNext());
+            assertEquals(centroid.count(), ix.next().count());
+        }
+        assertFalse(ix.hasNext());
+
+        buf.flip();
+        dist.asSmallBytes(buf);
+        assertTrue(buf.position() < 6000);
+        System.out.printf("# small %d bytes\n", buf.position());
+
+        buf.flip();
+        dist2 = fromBytes(buf);
+        assertEquals(dist.centroids().size(), dist2.centroids().size());
+        assertEquals(dist.compression(), dist2.compression(), 0);
+        assertEquals(dist.size(), dist2.size());
+
+        for (double q = 0; q < 1; q += 0.01) {
+            assertEquals(dist.quantile(q), dist2.quantile(q), 1e-6);
+        }
+
+        ix = dist2.centroids().iterator();
+        for (Centroid centroid : dist.centroids()) {
+            assertTrue(ix.hasNext());
+            assertEquals(centroid.count(), ix.next().count());
+        }
+        assertFalse(ix.hasNext());
+    }
+
+    @Test
+    public void testIntEncoding() {
+        Random gen = getRandom();
+        ByteBuffer buf = ByteBuffer.allocate(10000);
+        List<Integer> ref = Lists.newArrayList();
+        for (int i = 0; i < 3000; i++) {
+            int n = gen.nextInt();
+            n = n >>> (i / 100);
+            ref.add(n);
+            AbstractTDigest.encode(buf, n);
+        }
+
+        buf.flip();
+
+        for (int i = 0; i < 3000; i++) {
+            int n = AbstractTDigest.decode(buf);
+            assertEquals(String.format("%d:", i), ref.get(i).intValue(), n);
+        }
+    }
+
+    @Test
+    public void compareToQDigest() throws FileNotFoundException {
+        Random rand = getRandom();
+        PrintWriter out = new PrintWriter(new FileOutputStream("qd-tree-comparison.csv"));
+        try {
+            for (int i = 0; i < repeats(); i++) {
+                compareQD(out, new Gamma(0.1, 0.1, rand), "gamma", 1L << 48);
+                compareQD(out, new Uniform(0, 1, rand), "uniform", 1L << 48);
+            }
+        } finally {
+            out.close();
+        }
+    }
+
+    private void compareQD(PrintWriter out, AbstractContinousDistribution gen, String tag, long scale) {
+        for (double compression : new double[]{2, 5, 10, 20, 50, 100, 200, 500, 1000, 2000}) {
+            QDigest qd = new QDigest(compression);
+            TDigest dist = factory(compression).create();
+            List<Double> data = Lists.newArrayList();
+            for (int i = 0; i < 100000; i++) {
+                double x = gen.nextDouble();
+                dist.add(x);
+                qd.offer((long) (x * scale));
+                data.add(x);
+            }
+            dist.compress();
+            Collections.sort(data);
+
+            for (double q : new double[]{0.001, 0.01, 0.1, 0.2, 0.3, 0.5, 0.7, 0.8, 0.9, 0.99, 0.999}) {
+                double x1 = dist.quantile(q);
+                double x2 = (double) qd.getQuantile(q) / scale;
+                double e1 = cdf(x1, data) - q;
+                out.printf("%s\t%.0f\t%.8f\t%.10g\t%.10g\t%d\t%d\n", tag, compression, q, e1, cdf(x2, data) - q, dist.smallByteSize(), QDigest.serialize(qd).length);
+
+            }
+        }
+    }
+
+    @Test
+    public void compareToStreamingQuantile() throws FileNotFoundException {
+        Random rand = getRandom();
+
+        PrintWriter out = new PrintWriter(new FileOutputStream("sq-tree-comparison.csv"));
+        try {
+
+            for (int i = 0; i < repeats(); i++) {
+                compareSQ(out, new Gamma(0.1, 0.1, rand), "gamma", 1L << 48);
+                compareSQ(out, new Uniform(0, 1, rand), "uniform", 1L << 48);
+            }
+        } finally {
+            out.close();
+        }
+    }
+
+    private void compareSQ(PrintWriter out, AbstractContinousDistribution gen, String tag, long scale) {
+        double[] quantiles = {0.001, 0.01, 0.1, 0.2, 0.3, 0.5, 0.7, 0.8, 0.9, 0.99, 0.999};
+        for (double compression : new double[]{2, 5, 10, 20, 50, 100, 200, 500, 1000, 2000}) {
+            QuantileEstimator sq = new QuantileEstimator(1001);
+            TDigest dist = factory(compression).create();
+            List<Double> data = Lists.newArrayList();
+            for (int i = 0; i < 100000; i++) {
+                double x = gen.nextDouble();
+                dist.add(x);
+                sq.add(x);
+                data.add(x);
+            }
+            dist.compress();
+            Collections.sort(data);
+
+            List<Double> qz = sq.getQuantiles();
+            for (double q : quantiles) {
+                double x1 = dist.quantile(q);
+                double x2 = qz.get((int) (q * 1000 + 0.5));
+                double e1 = cdf(x1, data) - q;
+                double e2 = cdf(x2, data) - q;
+                out.printf("%s\t%.0f\t%.8f\t%.10g\t%.10g\t%d\t%d\n",
+                        tag, compression, q, e1, e2, dist.smallByteSize(), sq.serializedSize());
+
+            }
+        }
+    }
+
+    @Test()
+    public void testSizeControl() throws IOException, InterruptedException, ExecutionException {
+        // very slow running data generator.  Don't want to run this normally.  To run slow tests use
+        // mvn test -DrunSlowTests=true
+        assumeTrue(Boolean.parseBoolean(System.getProperty("runSlowTests")));
+
+        final Random gen0 = getRandom();
+        final PrintWriter out = new PrintWriter(new FileOutputStream("scaling.tsv"));
+        out.printf("k\tsamples\tcompression\tsize1\tsize2\n");
+
+        List<Callable<String>> tasks = Lists.newArrayList();
+        for (int k = 0; k < 20; k++) {
+            for (final int size : new int[]{10, 100, 1000, 10000}) {
+                final int currentK = k;
+                tasks.add(new Callable<String>() {
+                    Random gen = new Random(gen0.nextLong());
+
+                    @Override
+                    public String call() throws Exception {
+                        System.out.printf("Starting %d,%d\n", currentK, size);
+                        StringWriter s = new StringWriter();
+                        PrintWriter out = new PrintWriter(s);
+                        for (double compression : new double[]{2, 5, 10, 20, 50, 100, 200, 500, 1000}) {
+                            TDigest dist = factory(compression).create();
+                            for (int i = 0; i < size * 1000; i++) {
+                                dist.add(gen.nextDouble());
+                            }
+                            out.printf("%d\t%d\t%.0f\t%d\t%d\n", currentK, size, compression, dist.smallByteSize(), dist.byteSize());
+                            out.flush();
+                        }
+                        out.close();
+                        return s.toString();
+                    }
+                });
+            }
+        }
+
+        ExecutorService executor = Executors.newFixedThreadPool(20);
+        for (Future<String> result : executor.invokeAll(tasks)) {
+            out.write(result.get());
+        }
+        executor.shutdown();
+
+        out.close();
+    }
+
+    @Test
+    public void testScaling() throws FileNotFoundException, InterruptedException, ExecutionException {
+        final Random gen0 = getRandom();
+
+        PrintWriter out = new PrintWriter(new FileOutputStream("error-scaling.tsv"));
+        try {
+            out.printf("pass\tcompression\tq\terror\tsize\n");
+
+            Collection<Callable<String>> tasks = Lists.newArrayList();
+            int n = Math.max(3, repeats() * repeats());
+            for (int k = 0; k < n; k++) {
+                final int currentK = k;
+                tasks.add(new Callable<String>() {
+                    Random gen = new Random(gen0.nextLong());
+
+                    @Override
+                    public String call() throws Exception {
+                        System.out.printf("Start %d\n", currentK);
+                        StringWriter s = new StringWriter();
+                        PrintWriter out = new PrintWriter(s);
+
+                        List<Double> data = Lists.newArrayList();
+                        for (int i = 0; i < 100000; i++) {
+                            data.add(gen.nextDouble());
+                        }
+                        Collections.sort(data);
+
+                        for (double compression : new double[]{2, 5, 10, 20, 50, 100, 200, 500, 1000}) {
+                            TDigest dist = factory(compression).create();
+                            for (Double x : data) {
+                                dist.add(x);
+                            }
+                            dist.compress();
+
+                            for (double q : new double[]{0.001, 0.01, 0.1, 0.5}) {
+                                double estimate = dist.quantile(q);
+                                double actual = data.get((int) (q * data.size()));
+                                out.printf("%d\t%.0f\t%.3f\t%.9f\t%d\n", currentK, compression, q, estimate - actual, dist.byteSize());
+                                out.flush();
+                            }
+                        }
+                        out.close();
+                        System.out.printf("Finish %d\n", currentK);
+
+                        return s.toString();
+                    }
+                });
+            }
+
+            ExecutorService exec = Executors.newFixedThreadPool(16);
+            for (Future<String> result : exec.invokeAll(tasks)) {
+                out.write(result.get());
+            }
+            exec.shutdown();
+        } finally {
+            out.close();
+        }
+    }
+
+    @Test
+    public void testMerge() throws Exception {
+        merge(factory());
+    }
+
+    @Test
+    public void testExtremeQuantiles() {
+        // t-digest shouldn't merge extreme nodes, but let's still test how it would
+        // answer to extreme quantiles in that case ('extreme' in the sense that the
+        // quantile is either before the first node or after the last one)
+        TDigest digest = factory().create();
+        digest.add(10, 3);
+        digest.add(20, 1);
+        digest.add(40, 5);
+        // this group tree is roughly equivalent to the following sorted array:
+        // [ ?, 10, ?, 20, ?, ?, 50, ?, ? ]
+        // and we expect it to compute approximate missing values:
+        // [ 5, 10, 15, 20, 30, 40, 50, 60, 70]
+        List<Double> values = Arrays.asList(5., 10., 15., 20., 30., 40., 50., 60., 70.);
+        for (int i = 0; i < digest.size(); ++i) {
+            final double q = 1.0 / (digest.size() - 1); // a quantile that matches an array index
+            assertEquals(quantile(q, values), digest.quantile(q), 0.01);
         }
     }
 }
